# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Real-Time Messaging API",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "isActive": { "type": "boolean" },
          "lastOnlineAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Conversation": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "participants": { "type": "array", "items": { "type": "string" } },
          "type": { "type": "string", "enum": ["direct", "group"] },
          "lastMessageAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "conversationId": { "type": "string" },
          "senderId": { "type": "string" },
          "content": { "type": "string" },
          "status": { "type": "string", "enum": ["sent", "delivered", "read"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "deliveredAt": { "type": "string", "format": "date-time" },
          "readAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "username": {"type":"string"}, "email": {"type":"string"}, "password": {"type":"string"} }, "required": ["username","email","password"] } } }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "email": {"type":"string"}, "password": {"type":"string"} }, "required": ["email","password"] } } }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Refresh",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type":"object", "properties": { "refreshToken": {"type":"string"} }, "required":["refreshToken"] } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Logout",
        "requestBody": { "required": false, "content": { "application/json": { "schema": { "type":"object", "properties": { "refreshToken": {"type":"string"} } } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Me",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } } }
      }
    },
    "/api/user/list": {
      "get": {
        "summary": "List users",
        "security": [{ "BearerAuth": [] }],
        "parameters": [ {"name":"page","in":"query","schema":{"type":"integer"}}, {"name":"limit","in":"query","schema":{"type":"integer"}}, {"name":"q","in":"query","schema":{"type":"string"}} ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/conversations": {
      "get": {
        "summary": "List conversations",
        "security": [{ "BearerAuth": [] }],
        "parameters": [ {"name":"page","in":"query","schema":{"type":"integer"}}, {"name":"limit","in":"query","schema":{"type":"integer"}} ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/conversations/{id}": {
      "get": {
        "summary": "Conversation detail",
        "security": [{ "BearerAuth": [] }],
        "parameters": [ {"name":"id","in":"path","required":true,"schema":{"type":"string"}} ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/conversations/{id}/messages": {
      "get": {
        "summary": "List messages",
        "security": [{ "BearerAuth": [] }],
        "parameters": [ {"name":"id","in":"path","required":true,"schema":{"type":"string"}}, {"name":"page","in":"query","schema":{"type":"integer"}}, {"name":"limit","in":"query","schema":{"type":"integer"}}, {"name":"sort","in":"query","schema":{"type":"string"}}, {"name":"before","in":"query","schema":{"type":"string"}}, {"name":"after","in":"query","schema":{"type":"string"}} ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/messages": {
      "post": {
        "summary": "Create message",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "oneOf": [
            { "type":"object", "properties": { "conversationId": {"type":"string"}, "content": {"type":"string"} }, "required": ["conversationId","content"] },
            { "type":"object", "properties": { "toUserId": {"type":"string"}, "content": {"type":"string"} }, "required": ["toUserId","content"] }
          ] } } }
        },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } } } }
      }
    },
    "/api/messages/{id}/read": {
      "post": {
        "summary": "Read message",
        "security": [{ "BearerAuth": [] }],
        "parameters": [ {"name":"id","in":"path","required":true,"schema":{"type":"string"}} ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/online/count": {
      "get": { "summary": "Online count", "security": [{"BearerAuth":[]}], "responses": { "200": { "description": "OK" } } }
    },
    "/api/online/list": {
      "get": { "summary": "Online list", "security": [{"BearerAuth":[]}], "responses": { "200": { "description": "OK" } } }
    },
    "/api/search/messages": {
      "get": { "summary": "Search messages", "security": [{"BearerAuth":[]}], "parameters": [ {"name":"q","in":"query","required":true,"schema":{"type":"string"}}, {"name":"page","in":"query","schema":{"type":"integer"}}, {"name":"limit","in":"query","schema":{"type":"integer"}} ], "responses": { "200": { "description": "OK" } } }
    }
  }
}